conformance/extension/WEBCL_html_video/cl_commandQueue_enqueueWriteImage_with_HTMLVideoElement.html

<!--
 Copyright (C) 2013 Samsung Electronics Corporation. All rights reserved.

 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided the following condition
 are met:

 1.  Redistributions of source code must retain the above copyright
     notice, this list of conditions and the following disclaimer.

 2.  Redistributions in binary form must reproduce the above copyright
     notice, this list of conditions and the following disclaimer in the
     documentation and/or other materials provided with the distribution.

 THIS SOFTWARE IS PROVIDED BY SAMSUNG ELECTRONICS CORPORATION AND ITS
 CONTRIBUTORS "AS IS", AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING
 BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
 FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SAMSUNG
 ELECTRONICS CORPORATION OR ITS CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
 INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES(INCLUDING
 BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 DATA, OR PROFITS, OR BUSINESS INTERRUPTION), HOWEVER CAUSED AND ON ANY THEORY
 OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT(INCLUDING
 NEGLIGENCE OR OTHERWISE ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
 EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 -->
<html>
<head>
<meta charset="utf-8">
    <title>WebCLCommandQueue enqueueWriteImage with HTMLVideoElement conformance test.</title>
    <link rel="stylesheet" href="../../../resources/js-test-style.css"/>
    <script src="../../../resources/js-test-pre.js"></script>
    <script src="../../../resources/cl_utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
try {
    debug("WebCLCommandQueue enqueueWriteImage with HTMLVideoElement test case.");
    debug("(WebCLImage image, CLboolean blockingWrite, HTMLVideoElement srcVideo, optional sequence<WebCLEvent>? eventWaitList, optional WebCLEvent? event)\n");

    var wtu = WebCLTestUtils;
    var webCLContext = wtu.createContext();
    var webCLCommandQueue = wtu.createCommandQueue(webCLContext);

    var video = document.createElement('video');
    video.src = "../../../resources/red-green.mp4";
    video.setAttribute('type', "video/mp4");
    video.width = "320";
    video.height = "240";

    var webCLImage = webCLContext.createImage(webcl.MEM_READ_ONLY, video);
    shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(webCLImage, true, video);", "WEBCL_EXTENSION_NOT_ENABLED");
    if (!wtu.enableExtension(webcl, "WEBCL_html_video")) {
        throw { description : "Failed to enable extension : WEBCL_html_video. Exitting..."};
    }

    var webCLImageRead = webCLContext.createImage(webcl.MEM_READ_ONLY, video);
    shouldBeUndefined("webCLCommandQueue.enqueueWriteImage(webCLImageRead, true, video);");
    shouldBeUndefined("webCLCommandQueue.enqueueWriteImage(webCLImageRead, false, video);");

    // Testing for exceptions.
    shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(webCLImageRead, true, invalid_video);", "INVALID_HOST_PTR");

    var releasedImage = webCLContext.createImage(webcl.MEM_READ_ONLY, video);
    wtu.release(releasedImage);
    shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(releasedImage, true, video);", "INVALID_MEM_OBJECT");

    var webCLContextNew = wtu.createContext();
    var webCLCommandQueueNew = wtu.createCommandQueue(webCLContextNew);
    var webCLImageNew = webCLContextNew.createImage(webcl.MEM_READ_ONLY, video);
    shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(webCLImageRead, true, video);", "INVALID_CONTEXT");
    shouldThrowExceptionName("webCLCommandQueueNew.enqueueWriteImage(webCLImageRead, true, video);", "INVALID_CONTEXT");

    var webCLEvent = wtu.createEvent(webCLContext);
    var webCLEventNew = wtu.createEvent(webCLContextNew);
    shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(webCLImageRead, true, video, [webCLEventNew], webCLEvent);", "INVALID_CONTEXT");
    shouldThrowExceptionName("webCLCommandQueueNew.enqueueWriteImage(webCLImageRead, true, video, [webCLEvent], webCLEventNew);", "INVALID_CONTEXT");
    shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(webCLImageRead, true, video, [], webCLEventNew);", "INVALID_CONTEXT");
    shouldThrowExceptionName("webCLCommandQueueNew.enqueueWriteImage(webCLImageRead, true, video, [], webCLEvent);", "INVALID_CONTEXT");

    var releasedEvent = wtu.createEvent(webCLContext);
    shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(webCLImageRead, true, video, [releasedEvent], webCLEvent);", "INVALID_EVENT_WAIT_LIST");

    // FIXME To add test case to verify exception : INVALID_IMAGE_SIZE & INVALID_MEM_OBJECT.

} catch (e) {
    testFailed(e.description);
}
successfullyParsed = true;
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
